% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{boolean-class}
\alias{boolean-class}
\title{Boolean Model Class}
\value{
A boolean class contains the following items:
\describe{
\item{call}{call to estimate the model.}
\item{links}{link functions used to estimate the model.}
\item{model}{list, specifying the structure of the boolean model (used
             internally).}
\item{N}{number of observations.}
\item{k}{number of covariates.}
\item{coef.labels}{preformatted labels for the estimated coefficients.}
\item{coef.idx}{indicies for the coefficients in the model matrix (used
                internally).}
\item{response}{response vector.}
\item{frame}{model frame used to estimate the model.}
}
}
\description{
An object of class \code{boolean} as returned from \code{\link{boolprep}}.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

