% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{boolfit-class}
\alias{boolfit-class}
\title{A boolfit class}
\value{
A boolfit object is a component added to a
\code{\link{boolean-class}} object (as \code{model.fit}) once the model is
fit with \code{\link{boolean}}. It consists of a list of results returned by
the particular optimization routines used to fit the model.
}
\description{
An object of class boolfit.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

