% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{boolprob-class}
\alias{boolprob-class}
\title{A boolprob class}
\value{
The object returned by \code{\link{boolprob}}. It's comprised of the
following items:
\describe{
\item{est}{a data.frame of five columns: \code{llik}, the calculated range
           for the predicted probability
           \code{x}, the covariate value; \code{coef}, the
           preformatted variable name.}
\item{coef.labels}{preformatted coefficient labels.}
\item{default.plot}{a lattice plot produced by \code{xyplot}.}
}
}
\description{
An object of class boolprob.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

