% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{boolprof-class}
\alias{boolprof-class}
\title{A boolprof class}
\value{
The object returned by \code{\link{boolprof}}. It's comprised of the
following items:
\describe{
\item{est}{a data.frame of three columns: \code{llik}, the calculated
           log-likelihood; \code{x}, the coefficient value; \code{var}, the
           preformatted variable name.}
\item{coef.labels}{preformatted coefficient labels.}
\item{default.plot}{a lattice plot produced by \code{xyplot}.}
}
}
\description{
An object of class boolprof.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

