% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{boolsum-class}
\alias{boolsum-class}
\title{A boolsum class}
\value{
The object returned by \code{\link{summary.boolean}}.
}
\description{
An object of class boolsum.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

