% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coef.boolean}
\alias{coef.boolean}
\title{Return the model estimates.}
\usage{
\method{coef}{boolean}(object, ...)
}
\arguments{
\item{object}{boolean model object.}

\item{...}{Additional parameters to be passed.}
}
\value{
A vector of the coefficient estimates.
}
\description{
This function returns a vector of coefficient estimates.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

