% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logLik.boolean}
\alias{logLik.boolean}
\title{Return the log-likelihood for a boolean model.}
\usage{
\method{logLik}{boolean}(object, ...)
}
\arguments{
\item{object}{boolean model object.}

\item{...}{Additional parameters to be passed.}
}
\value{
This function returns an object of class \code{\link{logLik}}.
}
\description{
This function returns the log-likelihood of the estimated boolean model.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

