% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model.matrix.boolean}
\alias{model.matrix.boolean}
\title{Model Matrix from boolean object}
\usage{
\method{model.matrix}{boolean}(object, ...)
}
\arguments{
\item{object}{A boolean object.}

\item{...}{Arguments to pass on.}
}
\value{
An n-by-k model matrix.
}
\description{
Extract model matrix from boolean model.
}
\details{
This function extracts the model matrix from the specified boolean
model. Note that this model wouldn't always be appropriate for estimating a
model since multiple intercepts are included, making the columns of the
matrix perfectly collinear.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

