% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nobs.boolean}
\alias{nobs.boolean}
\title{Extract the Number of Observations from a Fit.}
\usage{
\method{nobs}{boolean}(object, ...)
}
\arguments{
\item{object}{boolean model object}

\item{...}{further arguments to be passed to other methods.}
}
\value{
An integer.
}
\description{
Extract the Number of Observations from a Fit.
}
\details{
Extract the Number of Observations from a Fit.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

