% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.boolprob}
\alias{plot.boolprob}
\title{Profile Predicted Probability Plot}
\usage{
\method{plot}{boolprob}(x, y = NULL, ...)
}
\arguments{
\item{x}{boolprob object.}

\item{y}{character or numeric vector specifying the covariates for which
the predicted probabilities should be plotted.}

\item{...}{additional arguments to pass to  \code{\link{update.trellis}}.
See \code{\link{xyplot}} for details.}
}
\value{
Plots the predicted probabilities for the selected covariates.
}
\description{
Plot predicted probabilities for selected covariates. Adjust default plot
produced by \code{\link{boolprob}} to fit the user's preferences.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

