% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.boolprof}
\alias{plot.boolprof}
\title{Profile Likelihood Plot}
\usage{
\method{plot}{boolprof}(x, y = NULL, ...)
}
\arguments{
\item{x}{boolprof object.}

\item{y}{character or numeric vector specifying the covariates for which
the profiled likelihood should be plotted.}

\item{...}{additional arguments to pass to  \code{\link{update.trellis}}.
See \code{\link{xyplot}} for details.}
}
\value{
Plots the profile likelihood for the selected covariates.
}
\description{
Plot the profile likelihoods for selected covariates. Adjust default plot
produced by \code{\link{boolprof}} to fit the user's preferences.
}
\author{
Jason W. Morgan <morgan.746@osu.edu>
}

