% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.boolean}
\alias{predict.boolean}
\title{Predicted Probabilities}
\usage{
\method{predict}{boolean}(object, ...)
}
\arguments{
\item{object}{boolean object.}

\item{...}{Additional parameters to pass on.}
}
\value{
A vector of predicted probabilities for the observations in the
model.
}
\description{
Calculate predicted probabilities
}
\details{
Calculate predicted probabilities for each observation.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

