% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.boolboot}
\alias{print.boolboot}
\title{Print Bootstrap Results for Boolean Object}
\usage{
\method{print}{boolboot}(x, ...)
}
\arguments{
\item{x}{boolboot object.}

\item{...}{Additional parameters to pass on.}
}
\value{
Print results in abbreviated summary.
}
\description{
Default print for boolboot objects.
}
\details{
Default print for boolboot objects.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

