% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.boolean}
\alias{print.boolean}
\title{Print a brief summary of a boolean model fit.}
\usage{
\method{print}{boolean}(x, ...)
}
\arguments{
\item{x}{boolean object.}

\item{...}{Additional parameters to be passed.}
}
\value{
print brief summary of the model and the model fit (if available).
}
\description{
This function prints a brief summary of the model fit to the user. The
details of a model fit are provided by \code{summary.boolean}. If the model
was not fit, this simply reports some summary statistics about the unfit
model.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

