% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.boolprob}
\alias{print.boolprob}
\title{Default print method for boolprob objects.}
\usage{
\method{print}{boolprob}(x, ...)
}
\arguments{
\item{x}{object of the \code{boolprob-class}.}

\item{...}{Additional arguments that are passed to the \code{lattice} plot
object.}
}
\value{
Plots the estimated predicted probabilities.
}
\description{
This is the default print object for \code{boolprob} objects.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

