% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.boolprof}
\alias{print.boolprof}
\title{Default print method for boolprof objects.}
\usage{
\method{print}{boolprof}(x, ...)
}
\arguments{
\item{x}{object of the \code{boolprof-class}.}

\item{...}{Additional arguments that are passed to the \code{lattice} plot
object.}
}
\value{
Plots the estimated log-likelihood profiles.
}
\description{
This is the default print object for \code{boolprof} objects.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

