% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.boolsum}
\alias{print.boolsum}
\title{Print summary of boolean model as described in \code{boolsum} object.}
\usage{
\method{print}{boolsum}(x, ...)
}
\arguments{
\item{x}{object of class \code{boolsum}.}

\item{...}{Additional parameters to be passed.}
}
\value{
Prints summary information about the boolean model fit.
}
\description{
This function prints a summary of a boolean model as described in a
\code{boolsum}, itself the result of calling \code{summary} on an object of
class \code{boolean}.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

