% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.boolboot}
\alias{summary.boolboot}
\title{Print Bootstrap Results for Boolean Object}
\usage{
\method{summary}{boolboot}(object, ...)
}
\arguments{
\item{object}{object of class boolboot.}

\item{...}{additional parameters to pass on.}
}
\value{
Prints a summary.
}
\description{
Summary function for boolboot objects.
}
\details{
Summary function for boolboot objects.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

