% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.boolean}
\alias{summary.boolean}
\title{Create a summary object from a boolean estimate.}
\usage{
\method{summary}{boolean}(object, ...)
}
\arguments{
\item{object}{object of class \code{boolean}.}

\item{...}{Additional parameters to be passed.}
}
\value{
A list of class \code{\link{boolsum-class}}.
}
\description{
This function calculates the standard model fit summary information given
the model object.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

