% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vcov.boolean}
\alias{vcov.boolean}
\title{Variance-Covariance Matrix for Boolean Model}
\usage{
\method{vcov}{boolean}(object, ...)
}
\arguments{
\item{object}{boolean object.}

\item{...}{Additional parameters to pass on.}
}
\value{
A matrix, the negative of the inverse Hessian.
}
\description{
Calculate the variance-covariance matrix for a boolean object.
}
\details{
Calculate the variance-covariance matrix for a boolean object.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

