% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.boolean}
\alias{predict.boolean}
\title{Predicted Probabilities}
\usage{
\method{predict}{boolean}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{boolean object.}

\item{newdata}{optionally, a data.frame containing the covariate profiles to
predict. If omitted, the original data used in fitting the model will be
used. Note that newdata must have the same structure as that returned by
model.matrix.}

\item{...}{Additional parameters to pass on.}
}
\value{
A vector of predicted probabilities for the covariate profiles
specified in newdata.
}
\description{
Calculate predicted probabilities
}
\details{
Calculate predicted probabilities for each observation.
}
\author{
Jason W. Morgan (\email{morgan.746@osu.edu})
}

